fixed behaviour of spotify's implicit grant flow redirect
app was unecessaryly redirecting since it already had an user access token. Added a condition at Spotify.getAccessToken to avoid it.
the current behaviour or redirect doesn't save whatever search input a user tried to search, however it would be wierder if the website asks for auth before loading
therefore, as a quickfix, getAccessToken() is called whenever a user type in for the first time.